requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# ML
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Data
yfinance>=0.2.0
fredapi>=0.5.0

# APIs
openai>=1.0.0
anthropic>=0.20.0
google-generativeai>=0.3.0

# Utils
python-dotenv>=1.0.0
pyyaml>=6.0
aiohttp>=3.9.0

# Viz
matplotlib>=3.7.0
seaborn>=0.12.0

# Test
pytest>=7.4.0
pytest-asyncio>=0.21.0
